backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.11.9
alembic==1.13.1
httpx==0.27.0
PyJWT==2.10.1
cryptography==44.0.3
numpy==2.4.2
scikit-learn==1.8.0
umap-learn==0.5.11
